Added tests for login action creators
