+++ priv/react_app/src/store/ActionCreators.js
@@ -36,46 +36,3 @@ export function getAllCompanies() {
 			})
 	};
 }
-
-export const loginFailure = () => {
-  return {
-    type: actions.LOGIN_FAILURE
-  }
-}
-
-export const loginSuccess = (jwt: string) => {
-  return {
-    type: actions.LOGIN_SUCCESS,
-    jwt
-  }
-}
-
-export const login = ({email, password}) => {
-  return dispatch => {
-    API.session.login({email, password})
-    .then(res => {
-      if(res.type === 'error') {
-        dispatch(loginFailure())
-      }
-      else {
-        const jwt = res.data.jwt
-        dispatch(loginSuccess(jwt))
-      }
-    })
-  }
-}
-
-export const development_login = (email) => {
-  return dispatch => {
-    API.session.development_login({email})
-    .then(res => {
-      if(res.type === 'error') {
-        dispatch(loginFailure())
-      }
-      else {
-        const jwt = res.data.jwt
-        dispatch(loginSuccess(jwt))
-      }
-    })
-  }
-}
