+++ priv/react_app/src/store/actions/Login/Login.js
@@ -1,0 +1,45 @@
+import actionTypes from './../../ActionTypes'
+import API from './../../../API'
+
+export const loginFailure = () => {
+  return {
+    type: actionTypes.LOGIN_FAILURE
+  }
+}
+
+export const loginSuccess = (jwt: string) => {
+  return {
+    type: actionTypes.LOGIN_SUCCESS,
+    jwt
+  }
+}
+
+export const login = ({email, password}) => {
+  return dispatch => {
+    API.session.login({email, password})
+    .then(res => {
+      if(res.type === 'error') {
+        dispatch(loginFailure())
+      }
+      else {
+        const jwt = res.data.jwt
+        dispatch(loginSuccess(jwt))
+      }
+    })
+  }
+}
+
+export const development_login = (email) => {
+  return dispatch => {
+    API.session.development_login({email})
+    .then(res => {
+      if(res.type === 'error') {
+        dispatch(loginFailure())
+      }
+      else {
+        const jwt = res.data.jwt
+        dispatch(loginSuccess(jwt))
+      }
+    })
+  }
+}
