+++ priv/react_app/src/store/actions/Login/index.js
@@ -1,0 +1,3 @@
+import * as Login from './Login'
+
+export default Login
