+++ priv/react_app/src/store/actions/index.js
@@ -1,0 +1,5 @@
+import login from './Login'
+
+export default {
+  login
+}
