+++ priv/react_app/src/store/test/TestHelper.js
@@ -1,0 +1,11 @@
+import configureMockStore from 'redux-mock-store'
+import thunk from 'redux-thunk'
+
+/**
+ * Creates a mockstore that can be used while testing
+ */
+const middlewares = [thunk]
+const createNewMockStore = configureMockStore(middlewares)
+export const createMockStore = (initialState) => {
+  return createNewMockStore(initialState)
+}
